+++ lib/webpack5/ChunkLoader.js
@@ -1,18 +1,17 @@
 // @ts-check
-const LoadScriptRuntimeModule = require('./LoadScriptRuntimeModule')
-const PublicPathRuntimeModule = require('./PublicPathModule')
-const AutoPublicPathRuntimeModule = require('./AutoPublicPathRuntimeModule')
+const LoadScriptRuntimeModule = require('./RuntimeModules/LoadScript')
+const PublicPathRuntimeModule = require('./RuntimeModules/PublicPath')
+const AutoPublicPathRuntimeModule = require('./RuntimeModules/AutoPublicPath')
+
 module.exports = class WebExtensionChuckLoaderRuntimePlugin {
-  /**
-   * @param {import('webpack').Compiler} compiler
-   */
+  /** @param {import('webpack').Compiler} compiler */
   apply(compiler) {
-    const RuntimeGlobals = compiler.webpack.RuntimeGlobals
+    const { RuntimeGlobals } = compiler.webpack
 
     compiler.hooks.compilation.tap(WebExtensionChuckLoaderRuntimePlugin.name, (compilation) => {
       compilation.hooks.runtimeRequirementInTree
         .for(RuntimeGlobals.loadScript)
-        .tap(WebExtensionChuckLoaderRuntimePlugin.name, (chunk, set) => {
+        .tap(WebExtensionChuckLoaderRuntimePlugin.name, (chunk) => {
           compilation.addRuntimeModule(
             chunk,
             LoadScriptRuntimeModule(
