+++ lib/webpack5/MV3ServiceWorkerPlugin.js
@@ -1,6 +1,6 @@
 // @ts-check
 const EagerlyLoadChunksRuntimeModule = require('./RuntimeModules/EagerlyLoadChunks')
-module.exports = class ManifestV3ServiceWorkerPlugin {
+module.exports = class MV3ServiceWorkerPlugin {
   /** @param {import('../..').BackgroundOptions} options */
   constructor(options) {
     this.options = options
@@ -8,7 +8,7 @@ module.exports = class ManifestV3ServiceWorkerPlugin {
   /** @param {import('webpack').Compiler} compiler */
   apply(compiler) {
     // Set chunkLoading to import-scripts
-    compiler.hooks.entryOption.tap(ManifestV3ServiceWorkerPlugin.name, (context, entries) => {
+    compiler.hooks.entryOption.tap(MV3ServiceWorkerPlugin.name, (context, entries) => {
       if (typeof entries === 'function') {
         if (this.options.noWarningDynamicEntry) return
         console.warn(`[webpack-extension-target] Dynamic entry points not supported yet.
@@ -27,8 +27,8 @@ Set serviceWorker.noWarningDynamicEntry to true to disable this warning.
     // Set all lazy chunks to eagerly loaded
     // See https://bugs.chromium.org/p/chromium/issues/detail?id=1198822
     if (this.options.eagerChunkLoading !== false) {
-      compiler.hooks.compilation.tap(ManifestV3ServiceWorkerPlugin.name, (compilation) => {
-        compilation.hooks.afterOptimizeChunkIds.tap(ManifestV3ServiceWorkerPlugin.name, () => {
+      compiler.hooks.compilation.tap(MV3ServiceWorkerPlugin.name, (compilation) => {
+        compilation.hooks.afterOptimizeChunkIds.tap(MV3ServiceWorkerPlugin.name, () => {
           const entryPoint = compilation.entrypoints.get(this.options.entry)
           if (!entryPoint) return
