+++ lib/webpack5/RuntimeModules/AutoPublicPath.js
@@ -1,5 +1,5 @@
 // @ts-check
-const BrowserRuntime = require('./BrowserRuntime')
+const BrowserRuntime = require('../BrowserRuntime')
 
 /**
  *
