+++ lib/webpack5/RuntimeModules/EagerlyLoadChunks.js
@@ -6,7 +6,7 @@ module.exports = function EagerlyLoadChunksRuntimeModule(webpack, chunks) {
   const { RuntimeGlobals, RuntimeModule, Template } = webpack
   class EagerlyLoadChunksRuntimeModule extends RuntimeModule {
     constructor() {
-      super('eagerly load chunks', 2000)
+      super('eagerly load chunks', RuntimeModule.STAGE_TRIGGER)
     }
     generate() {
       return Template.asString([chunks.map((x) => `${RuntimeGlobals.ensureChunk}(${JSON.stringify(x)});`)])
