+++ lib/webpack5/RuntimeModules/LoadScript.js
@@ -1,8 +1,7 @@
 // @ts-check
-const BrowserRuntime = require('./BrowserRuntime')
+const BrowserRuntime = require('../BrowserRuntime')
 
 /**
- *
  * @param {import('webpack')} webpack
  * @param {boolean | undefined} supportDynamicImport
  * @returns {import('webpack').RuntimeModule}
