+++ lib/webpack5/RuntimeModules/PublicPath.js
@@ -1,5 +1,5 @@
 // @ts-check
-const BrowserRuntime = require('./BrowserRuntime')
+const BrowserRuntime = require('../BrowserRuntime')
 
 /**
  * @param {import('webpack')} webpack
