+++ lib/webpack5/index.js
@@ -1,9 +1,13 @@
 // @ts-check
-const WebExtensionChuckLoaderRuntimePlugin = require('./RuntimePlugin')
-const NoDangerNamePlugin = require('./NoDangerNamePlugin')
-const ManifestV3ServiceWorkerPlugin = require('./ManifestV3ServiceWorkerPlugin')
 
-module.exports = class WebExtensionPlugin5 {
+// Make dynamic import & chunk splitting works.
+const ChuckLoaderRuntimePlugin = require('./ChunkLoader')
+// Ban invalid file names in web extension
+const NoDangerNamePlugin = require('./NoDangerNamePlugin')
+// Provide support for MV3
+const MV3ServiceWorkerPlugin = require('./MV3ServiceWorkerPlugin')
+
+module.exports = class WebExtensionPlugin {
   /**
    * @param {import('../../').WebExtensionPluginOptions} options
    */
@@ -16,10 +20,10 @@ module.exports = class WebExtensionPlugin5 {
   apply(compiler) {
     if (this.options.background) {
       if (this.options.background.manifest === 3) {
-        new ManifestV3ServiceWorkerPlugin(this.options.background).apply(compiler)
+        new MV3ServiceWorkerPlugin(this.options.background).apply(compiler)
       }
     }
-    new WebExtensionChuckLoaderRuntimePlugin().apply(compiler)
+    new ChuckLoaderRuntimePlugin().apply(compiler)
     new NoDangerNamePlugin().apply(compiler)
   }
 }
